
\name{balance}
\docType{data}
\alias{balance}

\title{Balance Scale Data}
\description{Balance scale data of five distance items from 472 subject.}
\usage{data(balance)}
\format{An object of class \code{markovdata}.}
\source{Jansen ... 2001}
\details{}
\references{Articles} 
\keyword{datasets}
