
\name{discrimination}
\docType{data}
\alias{discrimination}

\title{Discrimination Learning Data}
\description{

	This data set is from a simple discrimation learning experiment.  It
	consists of 192 binary series of responses of different lengths. This
	is a subset of the data described by \cite{Raijmakers et al (2001)}.
	
}
\usage{data(discrimination)}
\format{An object of class \code{markovdata}.}
\source{Maartje E. J. Raijmakers, Conor V. Doland and Peter C. M. Molenaar 
(2001).  Finite mixture distribution models of simple discrimination
learning. \emph{Memory \& Cognition}, vol 29(5).}

\keyword{datasets}
