
\name{mbd}
\docType{data}
\alias{mbd}

\title{Two Component Mixture Data Bentley}
\description{
	Data for illustration of fitting finite mixture distributions.
}
\usage{data(mbd)}
\format{An object of class \code{markovdata}.}
\source{M... & Bentley}
\details{}
\references{Articles} 
\keyword{datasets}
