
\name{discrimination}
\docType{data}
\alias{discrimination}

\title{Discrimination Learning Data}
\description{
	This data set is from a simple discrimation learning experiment.  It
	consists of 192 binary series of responses of different lengths. This
	is a subset of the data described by \cite{Raijmakers et al. (2001)}, and
	it is analyzed much more extensively using latent Markov models and depmix in
	\cite{Schmittmann et al. (2006)} and \cite{Visser et al. (2006).}.
}
\usage{data(discrimination)}
\format{An object of class \code{markovdata}.}
\source{
	Maartje E. J. Raijmakers, Conor V. Dolan and Peter C. M. Molenaar 
	(2001).  Finite mixture distribution models of simple discrimination
	learning. \emph{Memory \& Cognition}, vol 29(5).

	Ingmar Visser, Verena D. Schmittmann, and Maartje E. J. Raijmakers
	(2007).  Markov process models for discrimination learning.  In: Kees
	van Montfort, Han Oud, and Albert Satorra (Eds.), \emph{Longitudinal
	models in the behavioral and related sciences}, Mahwah (NJ): Lawrence
	Erlbaum Associates.

	Verena D. Schmittmann, Ingmar Visser and Maartje E. J. Raijmakers
	(2006).  Multiple learning modes in the development of rule-based
	category-learning task performance.  \emph{Neuropsychologia, vol
	44(11)}, p.  2079-2091.
}

\keyword{datasets}
